pyproject.toml
[tool.poetry]
name = "rag-performance"
version = "0.1.0"
description = "Benchmarking the performance of different RAG solutions"
authors = ["Nolan Tremelling <34580718+NolanTrem@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
haystack-ai = "^2.3.0"
langchain = "^0.2.8"
langchain-openai = "^0.1.16"
langchain-community = "^0.2.7"
langchain-chroma = "^0.1.2"
llama-index = "^0.10.55"
r2r = "^0.2.60"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true